[#128]Fix(반응하기): 반응하기 기능과 관련된 post TODO 코드들 추가 구현, post에 있던 comment 갯수 관련 기능 추가 구현

+++ src/main/java/today/sesac/versebyverse/comment/repository/CommentRepository.java
@@ -55,10 +55,10 @@ public interface CommentRepository extends JpaRepository<Comment, Long> {
     /**
      * 게시글 id로 해당 게시글의 삭제되지 않고, 신고되지 않은 댓글 수를 조회합니다.
      *
-     * @param id
+     * @param id 게시글 id
      * @return Long 으로 해당 게시글의 댓글 수
      */
-    Long countByPostIdAndIsDeletedFalseAndIsBlockedFalse(Long id);
+    int countByPostIdAndIsDeletedFalseAndIsBlockedFalse(Long id);
 
     /**
      * 게시글 id로 해당 게시글의 댓글 수를 조회합니다.
